Edited Don's code to be cleaner and added comments
Also the makefile works on the VM.
This currently nearly simulates a centralized algorithm solution to the problem, but doesn't utilize signals. Instead, it uses threads with a shared mutex lock associated with a shared chopstick array.
To improve, implement with signals so that all threads (including the leader) ask the leader thread (id = 0) if they can eat by sending a signal with their id. The leader should handle signals by replying with a signal of either yes or no (depending on if their dedicated chopsticks are available).
Then once a thread is done eating, it should send another type of signal similar to its first signal to signify they have returned the chopsticks.